requirements.txt
numpy>=1.24
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.5
drf-spectacular>=0.26.5
channels>=4.0
daphne>=4.0
asgiref>=3.7
httpx>=0.25
cryptography>=41.0
psycopg2-binary>=2.9
pytest>=7.4
pytest-django>=4.5
pytest-asyncio>=0.21
pytest-playwright>=0.4.0
black>=23.12
isort>=5.12
flake8>=6.1
mypy>=1.6
types-requests
types-python-dateutil
